Server: JWT Signing Algorithm Expansion (#994)
## Motivation

See #983. The JWT library that we use exposes a number of different
signing algorithms, and it'd be neat to allow people to use some of the
more common ones instead of the default HS256.

## Solution

This PR adds the `JwtSigningConfig` enum to the configuration which
wraps the algorithms exposed by `jwt-simple`. It's a fun, double-layered
enum to allow a fallback shape for backwards compatibility purposes.

The hard part is the deserialization. After that, the signing config
delegates generation and verification calls to whichever kind of key we
have loaded.

The signing config is wrapped in an `Arc` to avoid having many copies of
the key in memory, and also because several of the algorithms are not
`Clone`.

The non-`Default` variants are wrapped in a `Box` to avoid wasting a ton
of space when the `Default` variant is used (which is probably going to
be most of the time).

Finally, the signing config has a custom Debug impl that just spits out
the name of the algorithm in play, to avoid accidentally leaking the
key.