Bridge: better js transformations
This PR serves a few aims:

1. improve diagnostics around configuring transformations.
2. improve overall performance of transformations.
3. upgrade deno
3a. without the penalty of leaking memory.
3b. open the door to using async/await and JS module code (needed for
the "poller input" prototyped during our recent hack week.
3c. avoid that CVE we saw for all but recent deno versions.

This diff upgrades us from a very old `deno_core` to the current release
`0.195.0` via the latest `deno_runtime`. The latter is what provides
expanded JS features such as async/await and typical DOM APIs you might
expect to have, such as `fetch`.

In `deno_core` versions beyond `0.142.0`, we see memory leaked each time
we initialized a `JsRuntime` and drop it.

This was the old method for executing JS code in Bridge. With the upgrade, we
now mitigate the leak by retaining the `JsRuntime`s we make. A fixed
pool of `JsRuntime`s, each in a dedicated thread,  are initialized on
startup.
Permit-based accounting (via `deadpool`) is used to distribute work across
the pool.

With this new approach, casual load testing shows dramatically improved
throughput. Probably this is due to amortization of interpreter startup
time, which previously we paid for each and every message handled.

Casual load testing also showed the memory profile to be essentially
flat (~100Mb with 4 JS runtimes), but there's still the possibility of memory
leaks stemming from the JS code itself since the interpreters are now
long-living.

For the goal of better diagnostics, the config parser now passes each
transformation source string to `deno_ast` to check that the syntax is
at least valid. Since the config parser happens early on, this will mean
an error on startup instead of sometime later when an event tries to run
through the broken script.