Extend Rust enum generation template to handle integer-based enums (#1010)
## Motivation

We have a couple of integer-based enums in our API responses. The
existing template for enum generation in the Rust client library only
supports string-based enums.

Fixes #1009 

## Solution

We can switch on the type of enum as defined in the OpenAPI
specification. Then it's just a matter of pulling in serde_repr to do
the heavy lifting.