Server: hardcode irrelevant fields in the generated openapi (#1033)
We don't really care about the version nor the description when
generating libs, and both of them were increasing the diff unnecessarily
and making it annoying to merge things. We only use this file to make
sure that when we change the API we do it consciously, so these fields
don't matter.