Bridge: debulk consumer code
Essentially, once we've got omniqueue in bridge, we can "debulk" a lot
of the consumer code by un-traiting it.

A single concrete type can hold an input config which can be used to
produce a `DynConsumer` which can then be used to run the consumer loop.

This diff takes us _most of the way there_, keeping the trait, but now
building a single type which implements it from a range of config
values.