Remove ctx family of macros in favor of track_caller (#1054)
This PR alters the way we're doing error propagation and backtrace
handling.

Currently we use a family of macros that create and modify errors using
the `file!` and `line!` std macros. Our primary `ctx!` macro embeds the
location information into the call sites at compile time.

This PR instead piggybacks off of the `#[track_caller]` proc macro used
by rust's panic system. This embeds the caller information into the
binary in a similar way, but allows us to annotate a chain of functions
directly instead of a chain of macros. That means that we can do the
backtrace assembly automatically as part of the From impl, which means a
lot fewer manual tracing invocations.

I couldn't think of a way to automatically provide another trace line
when we're already propagating one of our own errors. This means that a
manual call to `trace()` is still required in a few places
unfortunately.

If there's a downside to this, I would say that reducing the tracing
calls (formerly ctx! and now trace()) from mandatory to sometimes means
that it's not always clear if it's necessary or desired to add one.