Set prerelease version before publishing to npm in Github action (#1076)
See
https://github.com/svix/svix-webhooks/actions/runs/6173804183/job/16756967707

Publishing the prerelease package failed because we didn't set the
correct version before `yarn publish`

## Solution
This uses `yarn version --prerelease --preid next`. For example,
`1.12.0` becomes `1.12.0-next.1` before publishing, so it doesn't
conflict with the current released version.

Also added a filter to only run the workflow if the javascript directory
was changed.