Simplify string caching
Since `StringCacheValue` is always a String, there's not much point in
having a separate type for it. This adds a marker trait that gets
implemented in the macro so that we can have still have type safety
around cache keys that store strings, but otherwise change the
relevant cache methods to just take/return `&str`/`String` as
appropriate.