[javascript] Fix no content-type error for empty responses (#1178)
In a previous PR https://github.com/svix/svix-webhooks/pull/1160 I added
an error message that is thrown before checking the response types, when
no content type is defined. That is causing issues in some cases since
there are cases in which the API returns no content type in the response
(for example, when the response code is `202 Accepted`).

Undoing that change for now. We need a better way to check for invalid
content-type responses, but I want to get this fix out first.