ci: Remove unnecessary extra cargo command args
For clippy and build, --all is deprecated in favor of --workspace, but is
also a no-op for workspaces unless some workspace members are not
default members.

For fmt, it already operates on the whole workspace by default; --all
just additionally includes non-workspace path dependencies (super rare
and should likely never be committed anyways).