Update redis health check (#1297)
Due to recent changes in the redis library, a PING command is now routed
to all main nodes by default, rather than a single random node, and all
must succeed. This makes connections particularly fragile if a single
node is having transient issues. So, in an effort to maximize uptime,
let's require a successful response from only one node.

This also updates omniqueue to the same version we use downstream. It
includes the same version of the redis lib/bb8.