Add NullableBool function (#1322)
<!--
Thank you for your Pull Request. Please provide a description above and
review
the requirements below.

Bug fixes and new features should include tests.
-->

## Motivation

The `AppPortalAccess` function in the golang SDK takes a `ReadOnly`
parameter, of type `NullableBool`. However, the `NullableBool` struct
and associated functions aren't accessible outside the SDK, so it's not
possible to construct a `NullableBool` to pass into this function.

For other types (`NullableString`, `NullableInt32`), there's a set of
wrapper functions defined in the `svix` package.

<!--
Explain the context and why you're making that change. What is the
problem
you're trying to solve? If a new feature is being added, describe the
intended
use case that feature fulfills.
-->

## Solution
Add a `NullableBool` function to the svix package, next to the existing
wrapper functions.

<!--
Summarize the solution and provide any necessary context needed to
understand
the code change.
-->