Libs(PHP): check timestamp is not float in `sign()`
We've got a unit test that checks to see if an exception is thrown when
the given timestamp is a float.

The test case (which is totally fair) uses an input ending in `.0` which
casts cleanly enough to an int with the existing "is_positive_integer"
check such that the exception isn't thrown.

Adding a call to `is_float()` to this check gets us back to the test
passing.

N.b. the `+ 0` there is to cast the `$timestamp` variable to an actual
numeric type instead of a string for the purpose of the call.