Bridge: add itertools for partition_map
OSS Review: itertools

This popular extension crate adds new methods for iterators.
Useful in this case as it cleans up a previous call to `fold()` designed
to partition a collection into 2 collections that can't be expressed
cleanly with a plain `partition()`.

The selected version is not the newest, but it aligns with a selected
version already included in the project transitively.