Bridge: factor backoffs to make sure we set sleep_time consistently
There were some gaps in the initial implementation. This switches to
setting a flag within the message handling loop and breaking the loop so
that the sleep time penalty (or reset) can be assessed at the bottom of
the loop more consistently.

Mostly this is about making sure any single message that has a problem
makes the whole iterator batch retry.