Bridge: use batch count=0 to signal backoff, not "done"
Looks like the `done` field doesn't signify we're at the end of the
topic like I originally thought.

Using the count as a proxy for this works well enough for now, but may
not in the future if we expose filters like channels/event types.

Follow-on from https://github.com/svix/svix-webhooks/pull/1363