Libs(Python): fix datetime serialization for `before`/`after
The API server expects timestamps to be formatted a specific way (an RFC
3339, UTC timestamp "with a Z").

This diff tweaks how we produce strings for the `before` and `after`
query params for message and attempt list requests, conforming to the
expected format.

For `before` and `after` we have a convenient place to influence
serialization by overriding the `to_dict` method, but a similar problem
exists exists for `since` and `until` used in `EndpointStatsOptions`.

I didn't see a trivial way to influence the serialization in these cases
without breaking the typing.

It seems like there should be a more uniform solution, but I don't yet
know what it is. Ideally all datetimes we pass through to the openapi
codegen API would be encoded the proper way.