Libs(Python): ensure datetimes have a TZ
The server fails to parse timestamps without any timezone.

This diff treats any timestamp without a TZ set as if it's in UTC.

Folks who want to pass a local time need to be sure to set the relevant
TZ themselves else we'll wind up treating it as UTC without shifting the
various components.