Rewrap doc comments to always have a succinct first paragraph
Clippy is unhappy about some of our doc comments, and I agree with it.
They're supposed to have a succinct first paragraph that can be rendered
as a summary (also helpful in our OpenAPI docs). This PR makes it so, at
least for all the comments that were surpassing the threshold that makes
clippy warn about an overly long first paragraph.