Add redis sentinel support
This adds support for using Redis Sentinel for queuing and caching.
Sentinel potentially requires a lot of additional configuration params,
so support for that that has been added here, although the `redis_dsn`
field has been reused and should point to a sentinel server if using the
`redissentinel` queue/cache types.