Libs(Go): add verbatim `model_simple.mustache` @ generator v5.2.0
We've got an issue with the codegen for optional/nullable fields in our
Go client when the field is an array.

A fix is coming in two revs. This rev adds the template source to our
repo as-is. The next will apply a workaround patch to correct the issue.

Note that our patches (in revs to follow) may need to be reapplied when
we update the version of the generator we target. For now, this aligns
with the specific version we're using.

Ref: <https://github.com/OpenAPITools/openapi-generator/blob/v5.2.0/modules/openapi-generator/src/main/resources/go/model_simple.mustache>