Libs(Go): update template, non-required nullable containers are pointers
Modifies the v5.2.0 `model_simple` template to make it so a non-required
nullable array such as `EndpointIn.channels` will be typed as a `*[]string`
instead of a `[]string`.

Previous versions of the generator did this, but later ones stopped for
reasons unknown.

Hand-edits to the `EndpointIn` model are faithfully reproduced by this
template update, but it's unclear if it'll cause issues elsewhere in the
codebase as written.

Existing Go tests are passing, and everything seems to typecheck with
the new codegen.