Libs(Go): get go tests running in CI
We've got a handful of unit test that could catch problems when the Go
sources are changing, so let's run them as a part of CI.

This doesn't re-run the OpenAPI codegen (but it could later). It would
be nice to know if the checked in sources were not aligned with whatever
the generator produces, for example if a recent spec bump PR neglected to
check them in.