Libs(Rust): update mode.mustache to generator `v7.9.0`
We had a local patch on an older version of this template to add support
for enums with integer reprs.

Seems like this is provided by default today, however the imports in the
default version of the template will cause a bunch of unused import
warnings.

This diff adds the v7.9.0 template as-is, and small mods will be made in
a future rev to fix the warnings.