Libs(Go): fresh (7.x) codegen, updates to high-level API
The codegen used here does not use the template overrides. There was a
healthy amount of symbol name changes, similar to those seen on the rust
side.

Tests are currently passing with some slight tweaks to account for the
template not being used (or some other disparity). Something about error
types are not the same as before. Will investigate.

Once cleared, and the templates are refactored, the test changes should
likely be rolled back.