Libs(Java): rework ApiClient template override for 7.x
In a previous rev, the template for ApiClient was reset to 7.9.0
(verbatim what the generator has bundled).

In this diff, we're adding back the retry support, but skipping the JSON
serializer customizations that used to be called from here.

In the upstream code, the JSON serializer was refactored considerably
and might even address the issues we aimed to fix in the first place by
manually editing the generated source.

For reference, those mods were made in https://github.com/svix/svix-webhooks/pull/1204

To start, leave the JSON customizations behind. We'll work to add some
tests to see how the client and serializer behave at 7.9.0, and rework
that patch if needed so it fits the updated upstream code.