Libs(Java): add kitchen sink test (like we have in Go)
This test is run against a local svix server to check some of the
behaviors around simple API interactions involving nullable fields.

Some refactors had to be made to the client initialization to allow the
us to override the server url deliberately. As things were, the url set
via SvixOptions would only be used when the region in the token was
unrecognized.

Just like with the Go kitchen sink test, it'll skip automatically when
the required env vars are not set. Put another way, you have to set the
following to run the test at all:

- `SVIX_TOKEN`
- `SVIX_SERVER_URL`