Libs(Go): update kitchen sink test to explicitly set descriptions
The description field for event type is required.

Seems like the "default zero value" for the string fields was
incorrectly accepted by the server.

Realized there was an issue when the server responded 422 on description-free
event types when I ported this test to Java. The Java models must not have
sent empty strings like Go might.