Libs(CSharp): fixup manual edits to -Patch models
We had some ignore patterns in place for a subset of -Patch models so
they weren't updated along with the rest of the models after the
generator was updated to 7.x.

For this diff I commented out the ignores, re-ran codegen, uncommented
the ignores again, then carefully restored the local edits. Mostly the
changes here are to set `EmitDefaultValue = false` for everything. This
should prevent model instances from supplying the default when a given
field isn't set (basically the same as `"omitempty"` on the Go side or
`skip_serializing_if = "Option::is_none"` in rust).

Looks like the generator will automatically assume this option should be
true for any container type, but does the thing we want (leave it false)
for scalars.

Note that by doing this, we're now filling in a handful of missing
fields and docstrings. We should figure out a better way to manage these
sorts of edits so we don't accidentally fail to track spec changes.