Libs(Kotlin): reapply local patches to 7.9.0 templates
Removes the enum_class mustache and custom serializer impl in the hopes
that the 7.9.0 generator handles these correctly without our patches.

Testing should confirm this. If they are still needed, we'll bring that
stuff back with the tests.

The ApiClient template had a block for running requests as a suspendable
coroutine which _should be present_ since we configure the generator
with `useCoroutines=true`, however I need to learn more kotlin before
I'm confident about how to fit the retry loop around that.

For now, I've left the blocking-style request we've always had in place
(i.e. the block you'd get when `useCoroutines=false`).

Parts of `build.gradle` have been updated to track the generator. The
kotlin version, for example, was needed to get the build working.

N.b. if the build starts hanging for you, clean out the `build` and
`generated` directories since there will be stale artifacts in each that
can interfere.