Libs(JS): reset template overrides at 7.9.0
Previously we had some manual edits made to `http/isomorphic-fetch`. In
order to smooth over future maintenance this diff adds the corresponding
mustache template for this file so we can track the upstream in a more
regular way.

Patches made to the `isomorphic-fetch` module will be applied to the
tempalte in the next rev, at which point it'll be removed from git. This
will allow me to see the diff plainly before/after the generator bump,
with the local patches factored in.

`MessageStatus` no longer seems to need a local edit since the new
generator handles int-repr enums correctly.

The `http/http` module didn't seem to have local edits at all.
I'm not sure why it was ignored/added to git in the first place.
Seems like it was added an ever modified.
The new generator's version of this module was quite different, so let's track
the upstream until a problem surfaces.

Anyway, the next rev will reapply local edits for `api` and cleanup the
rest of the above.