Libs(JS): add 7.9.0 generator template override for http/http
There were local edits made directly to this module prior to the bump.
The edits were not tracked in git, however. There looked to be just one
rev.

While testing things post-generator-bump, it looks like the edits made
were to help with the isomorphic fetch since I am seeing errors related
to symbols only available in browsers vs node.

This diff adds the template as-is, without patching. I'll rework the
template in a 2nd rev so we can actually see what the patch was.