Libs(JS): adding http to git again
This is an intermediate step while I troubleshoot some problems that
showed up during testing.

We previously tracked this module in git, but the local edits we made
were not tracked in history. I'll make sure to record whatever we change
this time, and move the edits into a template rather than a codegen
artifact.