Libs(JS): fixup http/http to avoid using `window`
Our JS client is meant to be suited for both browsers and node.
Looks like the only portion in the `http/http` module that doesn't work
in both is the `ensureAbsoluteUrl` function which would fallback to
using `window.location` as a prefix for the non-node case.

Instead, we'll just use the node behavior for everyone: throw an
exception if the URL isn't absolute.

Hopefully everything else lines up nicely after that.