Libs(JS): add "kitchen sink" test to show serialization works
I've been adding a test like this to each lib as I go while updating the
generator version. This should ensure the serialization is doing the
right thing with regards to nulls, and decoding response bodies even
while we're removing some of the patches we put in place to fix these
things.

Set `SVIX_TOKEN` and `SVIX_SERVER_URL` to run this against a live
`svix-server` instance.