Libs(Rust): add a "kitchen sink" test to verify null serialization
Just adding a little something to check how fields typed as nullable
collections are serialized.

New in this version of codegen are the "double Option" fields, allowing
us to set fields to null explicitly during PATCH requests!

I didn't test that, however. I tested that we're not sending "empty"
values when we shouldn't be since those would result in 422 responses
for clients who are otherwise following the rules.