Libs(Python): add kitchen sink test
Of the libs updated to support moving from the 5.2.0 openapi generator
up to 7.9.0, this one has sadly had the least attention (save for maybe
ruby).

For other libs, I've made an effort to update the dependencies to match
what's listed in the upstream generator, as well as "reset" the
override templates to match that in 7.9.0 then reapply our patches on
top of that. Normally I'd also try to remove templates that try to
address issues around body serialization.

Finally, I would add a small test to verify the behavior around nulls

I'm out of time, sadly so I'm trying to take a shortcut and (frankly)
hope for the best.

This diff adds the "kitchen sink" test to verify the nullables and so
on, but the templates and deps have been left as-is.

My hope is that even with the small collection of tests we've got, a
break from the template/generator/dependency mismatch would surface.

Likely, we can loop back and "modernize" this one.

The existing client tests had to be sort of disarmed to get the new test
to work.

setting `SVIX_TOKEN` and `SVIX_SERVER_URL` will avoid pytest trying to
launch docker containers, and also configure the kitchen sink to not be
skipped.

Example:
```
SVIX_TOKEN="xxx" SVIX_SERVER_URL="http://localhost:8071" pytest -k kitchen_sink
```