Revert "Libs(CSharp): add kitchen sink test"
This reverts commit 1f05d0eb8d2af3860762aa60c44f4f42c4512f24.

The test new runs fine on dotnet 8.0, but CI complains about something
in the source using language features too new.

The project currently targets dotnet 5.0, which seems to be EOL, so we
should look at updating the target first, after we can un-revert this
test.

For now, just know the test passed for me with a tweak to the Svix.Test
solution so it can run on 8.0. This gives me some confidence that the
generator bump produced some okay code.