Libs(Go): unestting nullable array fields via PATCH request
Work was done previously [1] to make nullable containers (arrays, maps)
generate as refs. This avoided a problem where the JSON serialization
code incrrectly sent an empty array for cases where a field was left
unset. This would result in requests incorrectly failing with a 422
response when it should have been OK.

During the openapi generator update [2], the resulting codegen no longer
had the 422 issue since it correctly omitted empty arrays from the body
for PATCH requests. However, this now means we have no way to explicitly
set a field to `null`. This specific issue was fixed in the some other langs
with the work done to update the generator, but the Go generator seems to
have fallen behind.

This rev adds tests to demonstrate the shortfall. We may need to rework
the template patch made for the old 5.x generator to solve this in the
meantime.

- 1: https://github.com/svix/svix-webhooks/pull/1450
- 2: https://github.com/svix/svix-webhooks/pull/1480