[Lib/JS] Compile svix package for ESM and CommonJS (#1549)
Closes https://github.com/svix/svix-webhooks/issues/1483

Compile the svix JavaScript package in `ESM` and `CommonJS`
simultaneously. This should solve the issues with tree-shaking (since it
can only be done when the package is compiled with ESM). Also keeps the
CommonJS version for backwards compatibility.

Tested this locally and all the usual imports (`import { EventTypeOut }
from "svix"`) work as expected, with no changes whatsoever.