Cli: add support for `listen` command
The `listen` command in the CLI allows you to use Svix Play as a public
webhook receiver, then via websocket, tunnel the webhooks through to
your local machine where they are forwarded to a local HTTP server.
Responses from the local server are shuttled back to Play via websocket
where they will show up in the UI.

This is a non-trivial change compared to the rest of the core CLI. The
translation from Go to Rust was not so straightforward, and not just
because of the difference in design of libs. Moving from a GC'd runtime
to a borrow-checked compliler means the original shape of the code is
not so easily reproducable.

Lots of manual testing indicates this _seems to work_ as expected.

The Go version has some color output support for the logging of messages
incoming and outgoing which are not here, but could be added later.

This is rough. My aim was to get something together that "works" so we
can move things forward. No doubt there are things that can be shifted
around to help organize this more idiomatically -- soon.