CLI: remove default for server url
The Go CLI treated the server url as a mandatory config value. Instead,
we're going to leave this unset when we instantiate a Svix client so
that it will use the server url of the token issuer.

The prompts for `svix login` now only ask for the auth token which
should cut down on the chance for a mismatch.

For the most part, server url is only something that should be
customized if targeting a non-production SaaS environment (dev, test,
staging and so on). This can still be configured by manually adding a
`server_url` entry to the `config.toml` or at invocation time via the
`SVIX_SERVER_URL`/`SVIX_DEBUG_URL` environment variables.