CLI: skip setting file mode for Windows
When the CLI creates a config file for the user, we use `600` unix
permissions to ensure only the user can read/write the file.

Setting the file mode like this is only possible like this on unix-like
systems so this breaks the build for Windows.

This diff splits the opening of the file into platform-specific
functions, one for `windows` which doesn't set the file mode.
The `unix` flavored function does what we had in the first place.

In <https://learn.microsoft.com/en-us/windows/win32/api/fileapi/nf-fileapi-createfile2>
there's mention of ACL being inherited from the parent directory. My
understanding is `%APPDATA%` should only be accessible by the named
user, so I'm hoping the defaults are correct for our use case.